Updated MovieTheatreGUI.java
Modified the class so that it is able to read the ArrayList of movies read in from the MovieText.txt file and convert the first *x* amount of movies into an array of movie names. I also modified the code so that it can efficiently save and restore the status of each seat depending on the colour.